+++ changelog.md
@@ -1,5 +1,30 @@
 ## Changelogs List
 
+## 0.0.3 - Altaev Nurzhan
+
+### About
+
+- Small Fixes for ContextMenu
+- Remove folder from ContextMenu(currently only on localeDB)
+
+### File Logs
+
+#### Added:
+
+- src/LeftPanel/components/LeftPanelItems/onContextMenu.tsx
+- src/components/ContextMenu/ContextItems.tsx
+- src/components/ContextMenu/ContextMenu.module.scss
+
+#### Changed:
+
+- api/index.cjs
+- api/surreal.cjs
+- electron-starter.babel.ts
+- src/LeftPanel/components/LeftPanelInput/LeftPanelInput.module.scss
+- src/LeftPanel/components/LeftPanelItems/LeftPanelItems.tsx
+- src/components/ContextMenu/ContextMenu.scss
+- src/components/ContextMenu/ContextMenu.tsx
+- src/redux/store/FoldersStore.tsx
 ## 0.0.2 - Altaev Nurzhan
 
 ### About